+++ database/migrations/1675007849227_request_items.ts
@@ -7,7 +7,7 @@ export default class extends BaseSchema {
     this.schema.createTable(this.tableName, (table) => {
       table.increments('id')
       table.string('name').notNullable()
-      table.integer('price').notNullable()
+      table.float('price').notNullable()
       table.integer('amount').notNullable()
       table
         .integer('request_id')
